requirements.txt
Flask==3.1.3
PyMySQL==1.1.1
python-dotenv==1.0.1
